Move tests to separate test suite
